pyproject.toml
[tool.poetry]
name = "nlpjunk"
version = "0.1.0"
description = ""
authors = ["sobamchan <oh.sore.sore.soutarou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pylint = "^2.6.0"
flake8 = "^3.8.4"
python-frontmatter = "^0.5.0"
mecab-python3 = "^1.0.3"
neologdn = "^0.4"
wordcloud = "^1.8.1"
fire = "^0.3.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-language-server = "^0.36.1"
ipython = "^7.19.0"
flake8 = "^3.8.4"
neovim = "^0.3.1"
jedi = "^0.17.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
